requirements.txt
groq
streamlit
requests
python-dotenv
pyyaml
pytest
# Optional: If you're using a specific search API
